+++ mteb/abstasks/Image/AbsTaskZeroShotClassification.py
