+++ mteb/abstasks/__init__.py
@@ -21,5 +21,5 @@ from .Image.AbsTaskImageClustering import *
 from .Image.AbsTaskImageMultilabelClassification import *
 from .Image.AbsTaskImageTextPairClassification import *
 from .Image.AbsTaskVisualSTS import *
-from .Image.AbsTaskZeroShotClassification_ import *
+from .Image.AbsTaskZeroShotClassification import *
 from .MultilingualTask import *
