+++ mteb/evaluation/evaluators/Image/ZeroShotClassificationEvaluator.py
