+++ mteb/evaluation/evaluators/__init__.py
@@ -10,7 +10,7 @@ from .Image.ClassificationEvaluator import *
 from .Image.ClusteringEvaluator import *
 from .Image.ImageTextPairClassificationEvaluator import *
 from .Image.VisualSTSEvaluator import *
-from .Image.ZeroShotClassificationEvaluator_ import *
+from .Image.ZeroShotClassificationEvaluator import *
 from .PairClassificationEvaluator import *
 from .RerankingEvaluator import *
 from .RetrievalEvaluator import *
