+++ mteb/tasks/Image/ZeroShotClassification/__init__.py
