+++ mteb/tasks/Image/ZeroShotClassification/eng/templates/Country211_labels.txt
