+++ mteb/tasks/Image/ZeroShotClassification/eng/templates/GTSRB_labels.txt
