+++ mteb/tasks/Image/ZeroShotClassification/eng/templates/Imagenet1k_labels.txt
