+++ mteb/tasks/Image/ZeroShotClassification/eng/templates/PatchCamelyon_labels.txt
